refactor(test-harness): summary rendering into render.rs
Pure restructure — the golden fixtures are unchanged.

The render_* functions, the per-suite table cells, the presentation caps
(MAX_LISTED, the compile-time thresholds), and percent()/commas() move to
summary/render.rs. Everything left in mod.rs is the public surface:
SummarySuite, render() orchestration, and the golden-fixture tests. The
module doc now names the layering: toolchain -> stats -> verdict ->
render.
